pyproject.toml
[tool.poetry]
name = "loadtest"
version = "0.1.0"
description = ""
authors = ["copyNdpaste <samnaka@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
locust = "^2.21.0"
black = "^24.1.1"
pymysql = "^1.1.0"
pytest = "^8.0.0"
python-dotenv = "^1.0.1"
pyjwt = "^2.8.0"
flask-jwt-extended = "^4.6.0"
sqlalchemy = "1.4.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
